#2 Refactor package naming from "api" and "resources"
